Move endpoint interface from endpoint to monitor
