+++ daemon/main.go
@@ -835,7 +835,7 @@ func runDaemon() {
 
 	repr, err := monitor.TimeRepr(time.Now())
 	if err != nil {
-		log.WithError(err).Warn("Failed to notify monitor about agent start")
+		log.WithError(err).Warn("Failed to generate agent start monitor message")
 	} else {
 		d.SendNotification(monitor.AgentNotifyStart, repr)
 	}
