+++ pkg/monitor/agent_test.go
@@ -17,6 +17,7 @@ package monitor
 import (
 	"encoding/json"
 	"fmt"
+	"sort"
 	"testing"
 	"time"
 
@@ -55,6 +56,9 @@ func testEqualityEndpoint(got, expected string, c *C) {
 	c.Assert(err, IsNil)
 	err = json.Unmarshal([]byte(expected), expectedStruct)
 	c.Assert(err, IsNil)
+
+	sort.Strings(gotStruct.Labels)
+	sort.Strings(expectedStruct.Labels)
 	c.Assert(gotStruct, comparator.DeepEquals, expectedStruct)
 }
 
@@ -83,7 +87,7 @@ func (s *MonitorSuite) TestRulesRepr(c *C) {
 	repr, err := PolicyUpdateRepr(rules, 1)
 
 	c.Assert(err, IsNil)
-	testEqualityRules(repr, "{\"labels\":[\"unspec:key1=value1\",\"unspec:key2=value2\"],\"revision\":1,\"rule_count\":2}", c)
+	testEqualityRules(repr, `{"labels":["unspec:key1=value1","unspec:key2=value2"],"revision":1,"rule_count":2}`, c)
 }
 
 func (s *MonitorSuite) TestRulesReprEmpty(c *C) {
@@ -92,7 +96,7 @@ func (s *MonitorSuite) TestRulesReprEmpty(c *C) {
 	repr, err := PolicyUpdateRepr(rules, 1)
 
 	c.Assert(err, IsNil)
-	testEqualityRules(repr, "{\"revision\":1,\"rule_count\":0}", c)
+	testEqualityRules(repr, `{"revision":1,"rule_count":0}`, c)
 }
 
 func (s *MonitorSuite) TestPolicyDeleteRepr(c *C) {
@@ -106,7 +110,7 @@ func (s *MonitorSuite) TestPolicyDeleteRepr(c *C) {
 
 	repr, err := PolicyDeleteRepr(1, lab.GetModel(), 2)
 	c.Assert(err, IsNil)
-	testEqualityRules(repr, "{\"labels\":[\"unspec:key1=value1\"],\"revision\":2,\"rule_count\":1}", c)
+	testEqualityRules(repr, `{"labels":["unspec:key1=value1"],"revision":2,"rule_count":1}`, c)
 }
 
 type RegenError struct{}
@@ -141,11 +145,11 @@ func (s *MonitorSuite) TestEndpointRegenRepr(c *C) {
 
 	repr, err := EndpointRegenRepr(e, rerr)
 	c.Assert(err, IsNil)
-	testEqualityEndpoint(repr, "{\"id\":10,\"labels\":[\"unspec:key1=value1\",\"unspec:key2=value2\"],\"error\":\"RegenError\"}", c)
+	testEqualityEndpoint(repr, `{"id":10,"labels":["unspec:key1=value1","unspec:key2=value2"],"error":"RegenError"}`, c)
 
 	repr, err = EndpointRegenRepr(e, nil)
 	c.Assert(err, IsNil)
-	testEqualityEndpoint(repr, "{\"id\":10,\"labels\":[\"unspec:key1=value1\",\"unspec:key2=value2\"]}", c)
+	testEqualityEndpoint(repr, `{"id":10,"labels":["unspec:key1=value1","unspec:key2=value2"]}`, c)
 }
 
 func (s *MonitorSuite) TestTimeRepr(c *C) {
@@ -154,5 +158,5 @@ func (s *MonitorSuite) TestTimeRepr(c *C) {
 	repr, err := TimeRepr(t)
 
 	c.Assert(err, IsNil)
-	c.Assert(repr, Equals, fmt.Sprintf("{\"time\":\"%s\"}", t.String()))
+	c.Assert(repr, Equals, fmt.Sprintf(`{"time":"%s"}`, t.String()))
 }
